feat!: Separate extractors and seasonality (#113)
Also speed up Lempel Ziv
