+++ docs/ref/feature-extraction.md
@@ -1,5 +1,3 @@
-::: functime.feature_extraction
-
 ## Time-series Feature Extractors
 
-:::functime.feature_extraction.tsfresh
+:::functime.feature_extractors
