+++ docs/ref/seasonality.md
@@ -1,0 +1,3 @@
+## Seasonality and Holiday Effects
+
+:::functime.seasonality
